+++ lib/moxl/src/Xec/Payload/Location.php
@@ -10,11 +10,18 @@ class Location extends Payload
     {
         $from = current(explode('/', (string)$parent->attributes()->from));
 
-        if (isset($stanza->items->item->geoloc)
-        && $stanza->items->item->geoloc->count() > 0) {
+        if (isset($stanza->items->item->geoloc)) {
             $contact = Contact::firstOrNew(['id' => $from]);
-            $contact->setLocation($stanza);
+            $contact->setLocation($stanza->items->item);
             $contact->save();
+
+            if ($from == \App\User::me()->id) {
+                $this->event('mylocation');
+                $this->event('mypresence');
+            } else {
+                $this->pack($contact);
+                $this->deliver();
+            }
         }
     }
 }